requirements.txt
# PlantMD Pro — Backend Dependencies
# Pin versions for reproducibility

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# ── ML / Deep Learning ───────────────────────────────────────────────────────
tensorflow==2.16.1          # Core DL framework
numpy==1.26.4
Pillow==10.3.0              # Image I/O
opencv-python-headless==4.9.0.80  # Grad-CAM overlay

# ── Configuration & Validation ───────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Caching ───────────────────────────────────────────────────────────────────
redis[asyncio]==5.0.4

# ── Logging ───────────────────────────────────────────────────────────────────
structlog==24.1.0

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit==1.35.0
requests==2.32.2

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0               # For async test client
factory-boy==3.3.0

# ── Code Quality ─────────────────────────────────────────────────────────────
black==24.4.2
isort==5.13.2
flake8==7.0.0
mypy==1.10.0
pre-commit==3.7.1
